Ensure that $canonical_self, when inspected, still has unmerged data
$canonical_self->SetComments() calls a $self->Load( $self->id ) during
the process of updating the Comments.  Because this now loads the data
of the user we are merged into, the information recorded on $merged was
always contentless -- namely, it always showed the the user had been
merged into itself.

Swap the order of the ->SetComments calls, to ensure that
$canonical_self is still the user being merged in when it is inspected
to record data on $merge.