->CanonicalizeEmailAddress may be a class method; don't assume $self
03573403 switched several uses of $RT::SystemUser to $self->CurrentUser,
on the basis that the current user should be propagated throughout,
rather than fixed to the current user.  However, the change in
CanonicalizeEmailAddress was overzealous, as that can be called as a
class method.

As the $canonical_user in this context is lexical, is not altered, and
should not be ACL'd in any way, hardcoding the system user is correct;
restore it.