Confirm loading effective id, return original on failure
When running shredder to clean up users, the effective id
for a merged ticket can be shredded. Add error checking
when loading an effective id to detect this case. On
failure, return the id of the current user id and warn
that the user it was merged into is missing.