Support to delete group members for merged users
As RT::Group::UserMembersObj returns ids of main users(i.e. users that
others are merged into), these main ids are passed on member delete on
admin group members page, no matter if it's merged users or the main
user that got added to the group.

Thus we need to find out all the merged users in DeleteMember and delete
them accordingly.