Do not try to load user from an empty email address
It's possible that DB finds a user record that doesn't have email
address. If the user was merged into another record that does have email
address, this unrelated email address would be returned as the canonical
value of empty email, which wrongly prevented RT from creating users
with empty emails.