Do not merge users if their Privileged values are different
When changing the value of a user from Privileged or UnPrivileged RT
does a HasMember lookup. If we allow Privileged and UnPrivileged users
to be merged, it means we can no longer edit this attribute because RT
will always think the user is a member of the Privileged or UnPrivileged
group when attempting to update the Privileged value. Resulting in no
change and a message like "Privileged status: That user is already unprivileged".